requirements.txt
locust>=2.29
pytest>=8.2
pytest-xdist>=3.6
httpx>=0.27
pydantic>=2.8
pydantic-settings>=2.5
pyyaml>=6.0
faker>=26.0
jinja2>=3.1
click>=8.1